# (#3921) - speed up updateDoc()