--- a/lib/adapters/idb/bulkDocs.js
+++ b/lib/adapters/idb/bulkDocs.js
@@ -193,6 +193,9 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
 
     docCountDelta += delta;
 
+    docInfo.metadata.winningRev = winningRev;
+    docInfo.metadata.deleted = winningRevIsDeleted;
+
     var doc = docInfo.data;
     doc._id = docInfo.metadata.id;
     doc._rev = docInfo.metadata.rev;