--- a/lib/deps/docs/updateDoc.js
+++ b/lib/deps/docs/updateDoc.js
@@ -5,28 +5,19 @@ var errors = require('../errors');
 var isDeleted = require('./isDeleted');
 var parseDoc = require('./parseDoc').parseDoc;
 
-function revExists(metadata, rev) {
-  var found = false;
-  merge.traverseRevTree(metadata.rev_tree, function (leaf, pos, id) {
-    if ((pos + '-' + id) === rev) {
-      found = true;
-    }
-  });
-  return found;
-}
-
 function updateDoc(prev, docInfo, results, i, cb, writeDoc, newEdits) {
 
-  if (revExists(prev, docInfo.metadata.rev)) {
+  if (merge.revExists(prev.rev_tree, docInfo.metadata.rev)) {
     results[i] = docInfo;
     return cb();
   }
 
-  // TODO: some of these can be pre-calculated, but it's safer to just
-  // call merge.winningRev() and isDeleted() all over again
-  var previousWinningRev = merge.winningRev(prev);
-  var previouslyDeleted = isDeleted(prev, previousWinningRev);
-  var deleted = isDeleted(docInfo.metadata);
+  // sometimes this is pre-calculated. historically not always
+  var previousWinningRev = prev.winningRev || merge.winningRev(prev);
+  var previouslyDeleted = 'deleted' in prev ? prev.deleted :
+    isDeleted(prev, previousWinningRev);
+  var deleted = 'deleted' in docInfo.metadata ? docInfo.metadata.deleted :
+    isDeleted(docInfo.metadata);
   var isRoot = /^1-/.test(docInfo.metadata.rev);
 
   if (previouslyDeleted && !deleted && newEdits && isRoot) {
@@ -64,7 +55,14 @@ function updateDoc(prev, docInfo, results, i, cb, writeDoc, newEdits) {
   var delta = (previouslyDeleted === winningRevIsDeleted) ? 0 :
     previouslyDeleted < winningRevIsDeleted ? -1 : 1;
 
-  var newRevIsDeleted = isDeleted(docInfo.metadata, newRev);
+  var newRevIsDeleted;
+  if (newRev === winningRev) {
+    // if the new rev is the same as the winning rev, we can reuse that value
+    newRevIsDeleted = winningRevIsDeleted;
+  } else {
+    // if they're not the same, then we need to recalculate
+    newRevIsDeleted = isDeleted(docInfo.metadata, newRev);
+  }
 
   writeDoc(docInfo, winningRev, winningRevIsDeleted, newRevIsDeleted,
     true, delta, i, cb);