--- a/tests/integration/index.html
+++ b/tests/integration/index.html
@@ -28,6 +28,7 @@
     <script src='deps/pouchdb-2.2.0-postfixed.js'></script>
     <script src='deps/pouchdb-3.0.6-postfixed.js'></script>
     <script src='deps/pouchdb-3.2.0-postfixed.js'></script>
+    <script src='deps/pouchdb-3.6.0-postfixed.js'></script>
     <script src='utils.js'></script>
     <script src='test.aa.setup.js'></script>
     <script src='test.basics.js'></script>