For now, I just want to see if the tests are passing.

This is a slightly dangerous optimization and may require a migration.
The fact that it seems to be working without migrations kinda worries
me; I think I might need to add a new migration test to catch it.

Also I have not measured the performance boost yet, but this is
pre-caching a lot of values as well as speeding up `revExists()`.