@nolanlawson 2015-08-06
I'm stunned this actually passed. This weekend I'll get some numbers on the perf boost and try to confirm whether or not we need a migration.