@nolanlawson 2015-08-09
I've added a migration test, so I feel confident now that this PR is good to go.

Speedup is slight, but any speedup is good, right? Numbers are from a low-spec Acer Aspire laptop running Linux Mint for the `bulk-inserts` test:

Chrome 11 (websql): 33002ms -> 32277ms
Chrome 44 (idb): 15690ms -> 14988ms
iojs 2.5 (leveldb): 4878ms -> 4862md