@daleharvey 2015-08-10
The failure is just a coveralls issue that seems to happen on branches that arent completely up to date
+1 on extra migration tests - https://github.com/pouchdb/pouchdb/commit/cd1357a6a63a048a71336a336e516a3d02afad6e